+++ tuna/miopen/db/miopen_tables.py
@@ -27,10 +27,10 @@
 """ Module for creating DB tables"""
 import enum
 from sqlalchemy import Column, Integer, String, UniqueConstraint, ForeignKey, DateTime
-from sqlalchemy import Text, Enum, Index
+from sqlalchemy import Enum, Index
 from sqlalchemy import Float, BigInteger, Boolean
 from sqlalchemy.databases import mysql
-from sqlalchemy.dialects.mysql import TINYINT, DOUBLE, MEDIUMBLOB, LONGBLOB
+from sqlalchemy.dialects.mysql import TINYINT, MEDIUMBLOB, LONGBLOB
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func as sqla_func
 from sqlalchemy.ext.declarative import declared_attr
@@ -42,6 +42,8 @@ from tuna.miopen.utils.config_type import ConfigType
 from tuna.miopen.db.session import Session
 from tuna.miopen.utils.metadata import DIR_MAP
 from tuna.miopen.db.benchmark import Model, Framework
+from tuna.db.tuna_tables import JobMixin
+from tuna.db.tuna_tables import FinStep
 
 COMMON_UNIQ_FDS = ["config", "solver", "session"]
 
@@ -307,6 +309,22 @@ class FusionConfigTags(BASE, ConfigTagMixin):
   config = Column(Integer, ForeignKey("fusion_config.id"), nullable=False)
 
 
+class MiopenJob(JobMixin):
+  """Represents Mixin class for job tables"""
+
+  compile_start = Column(DateTime,
+                         nullable=False,
+                         server_default=sqla_func.now())
+  compile_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
+  eval_start = Column(DateTime, nullable=False, server_default=sqla_func.now())
+  eval_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
+
+  solver = Column(String(length=128), nullable=True, server_default="")
+  fin_step = Column(mysql.MSSet(*(list(k for k in FinStep.__members__))),
+                    nullable=False,
+                    server_default="not_fin")
+
+
 class JobEnum(enum.Enum):
   """Represents job_enum column in config table"""
   # pylint: disable=invalid-name ; names represent entries in job_enum column
@@ -335,51 +353,6 @@ class JobEnum(enum.Enum):
   not_tunable = 22
   compiled_pend = 23
   evaluated_pend = 24
-
-
-class FinStep(enum.Enum):
-  """ Allowed Fin Steps """
-  # pylint: disable=invalid-name ; tuna/go_fish.py names valid fin steps as FinStep.__members__
-  find_compile = 1
-  find_eval = 2
-  get_solvers = 3
-  get_applicability = 4
-  not_fin = 5
-  miopen_find_compile = 6
-  miopen_find_eval = 7
-  miopen_perf_compile = 8
-  miopen_perf_eval = 9
-
-
-class JobMixin():
-  """Represents Mixin class for job tables"""
-
-  @declared_attr
-  def session(self):
-    """session key"""
-    return Column(Integer, ForeignKey("session.id"), nullable=False)
-
-  reason = Column(String(length=60), nullable=False, server_default="")
-  state = Column(Enum(JobEnum), nullable=False, server_default="new")
-  retries = Column(Integer, nullable=False, server_default="0")
-  result = Column(Text, nullable=True)
-
-  compile_start = Column(DateTime,
-                         nullable=False,
-                         server_default=sqla_func.now())
-  compile_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
-  eval_start = Column(DateTime, nullable=False, server_default=sqla_func.now())
-  eval_end = Column(DateTime, nullable=False, server_default=sqla_func.now())
-
-  gpu_id = Column(Integer, nullable=False, server_default="-1")
-  machine_id = Column(Integer, nullable=False, server_default="-1")
-  eval_mid = Column(Integer, server_default="-1")
-  cache_loc = Column(Text)
-
-  solver = Column(String(length=128), nullable=True, server_default="")
-  fin_step = Column(mysql.MSSet(*(list(k for k in FinStep.__members__))),
-                    nullable=False,
-                    server_default="not_fin")
 
 
 class ConvolutionJob(BASE, JobMixin):
