+++ vars/utils.groovy
@@ -608,7 +608,8 @@ def runLint() {
             sh "yamllint tuna/example/*.yaml"
             sh "mypy tuna/miopen/driver/base.py --ignore-missing-imports --follow-imports=skip"
             sh "mypy tuna/machine.py --ignore-missing-imports --follow-imports=skip"
-            sh "mypy tuna/session_mixin.py --ignore-missing-imports --follow-imports=skip"
+            sh "mypy tuna/db/session_mixin.py --ignore-missing-imports --follow-imports=skip"
+            sh "mypy tuna/db/tuna_tables.py --ignore-missing-imports --follow-imports=skip"
             sh "mypy tuna/parse_args.py --ignore-missing-imports --follow-imports=skip"
             sh "mypy tuna/worker_interface.py --ignore-missing-imports --follow-imports=skip"
             sh "mypy tuna/tables_interface.py --ignore-missing-imports --follow-imports=skip"
